Seperate client and server
