--- client/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "client",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "apollo-boost": "^0.1.16",
-    "graphql": "^14.0.2",
-    "react": "^16.5.1",
-    "react-apollo": "^2.1.11",
-    "react-dom": "^16.5.1",
-    "react-scripts": "1.1.5"
-  },
-  "proxy": "http://localhost:4000/",
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  },
-  "devDependencies": {
-    "eslint-plugin-graphql": "^2.1.1"
-  }
-}
