+++ package.json
@@ -1,46 +1,67 @@
 {
-    "name": "y",
-    "version": "1.0.0",
-    "description": "",
-    "main": "index.js",
-    "scripts": {
-        "start": "nodemon --inspect --ignore client/ -e js,graphql index.js",
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "lint:gql": "npx graphql-schema-linter schema.graphql"
+  "name": "client",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "client": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
+    "server": "nodemon --inspect --ignore src/ -e js,graphql server/apollo-server.js",
+    "lint:all": "eslint --fix --ignore-pattern 'node_modules/' .",
+    "lint:schema": "npx graphql-schema-linter schema.graphql"
+  },
+  "proxy": "http://localhost:4000/",
+  "graphql-schema-linter": {
+    "rules": [
+      "defined-types-are-used",
+      "deprecations-have-a-reason",
+      "enum-values-all-caps",
+      "enum-values-sorted-alphabetically",
+      "fields-are-camel-cased",
+      "input-object-values-are-camel-cased",
+      "relay-connection-arguments-spec",
+      "relay-connection-types-spec",
+      "relay-page-info-spec",
+      "types-are-capitalized"
+    ]
+  },
+  "dependencies": {
+    "apollo-boost": "^0.1.16",
+    "apollo-cache-inmemory": "^1.2.10",
+    "apollo-link-batch-http": "^1.2.3",
+    "apollo-server-express": "^2.0.7",
+    "express": "^4.16.3",
+    "graphql": "^14.0.2",
+    "react": "^16.5.2",
+    "react-apollo": "^2.1.11",
+    "react-dom": "^16.5.2",
+    "react-scripts": "2.0.5"
+  },
+  "devDependencies": {
+    "eslint-config-prettier": "^3.1.0",
+    "eslint-plugin-prettier": "^3.0.0",
+    "graphql-schema-linter": "^0.1.6",
+    "nodemon": "^1.18.4",
+    "prettier": "^1.14.3"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "prettier"
+    ],
+    "rules": {
+      "func-names": "off",
+      "prettier/prettier": "error"
     },
-    "author": "",
-    "license": "ISC",
-    "graphql-schema-linter": {
-        "rules": [
-            "defined-types-are-used",
-            "deprecations-have-a-reason",
-            "enum-values-all-caps",
-            "enum-values-sorted-alphabetically",
-            "fields-are-camel-cased",
-            "input-object-values-are-camel-cased",
-            "relay-connection-arguments-spec",
-            "relay-connection-types-spec",
-            "relay-page-info-spec",
-            "types-are-capitalized"
-        ]
-    },
-    "dependencies": {
-        "apollo-cache-inmemory": "^1.2.10",
-        "apollo-link-batch-http": "^1.2.3",
-        "apollo-server-express": "^2.0.7",
-        "express": "^4.16.3",
-        "graphql": "^14.0.2"
-    },
-    "devDependencies": {
-        "eslint": "^5.3.0",
-        "eslint-config-airbnb": "17.1.0",
-        "eslint-config-prettier": "^3.0.1",
-        "eslint-plugin-import": "^2.14.0",
-        "eslint-plugin-jsx-a11y": "^6.1.1",
-        "eslint-plugin-prettier": "^2.6.2",
-        "eslint-plugin-react": "^7.11.0",
-        "graphql-schema-linter": "^0.1.6",
-        "nodemon": "^1.18.4",
-        "prettier": "^1.14.2"
-    }
+    "plugins": [
+      "prettier"
+    ]
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
