+++ src/app/App.jsx
@@ -1,9 +1,8 @@
 import React, { Component } from "react";
 
-import logo from "./logo.svg";
-import ApolloPiece from "./ApolloPiece";
+import logo from "../logo.svg";
+import Rob from "./Rob";
 import Hider from "./Hider";
-import "./App.css";
 
 class App extends Component {
   render() {
@@ -17,7 +16,7 @@ class App extends Component {
           To get started, edit <code>src/App.js</code> and save to reload.
         </p>
         <Hider show={true}>
-          <ApolloPiece />
+          <Rob />
         </Hider>
       </div>
     );
