+++ src/app/Rob.jsx
@@ -1,35 +1,34 @@
 import React from "react";
 import { Query } from "react-apollo";
 import gql from "graphql-tag";
-import Hider from './Hider'
-import Terry from './Terry'
+import Hider from "./Hider";
+import Terry from "./Terry";
 
-export const mainQuery = gql`
-  query MainQuery($idPersons: ID!, $user_token: ID!) {
-    whoami(id: $user_token) {
-      id
-      name
-    }
-    persons(id: $idPersons) {
-      id
-      age
-      name
-      luckLevel
-    }
-  }
-`;
-export default function ApolloPiece({ bG, queryInfo, children }) {
+export default function Rob() {
   return (
     <Query
-      query={mainQuery}
-      variables={{
-        idPersons: "1234",
-        user_token: "890"
-      }}
+      variables={{ idPersons: "1234", user_token: "890" }}
+      query={gql`
+        query MainQuery($idPersons: ID!, $user_token: ID!) {
+          whoami(id: $user_token) {
+            id
+            name
+          }
+          persons(id: $idPersons) {
+            id
+            age
+            name
+            luckLevel
+          }
+        }
+      `}
     >
       {({ loading, error, data }) => {
         if (loading) return <p>Loading...</p>;
-        if (error) return <p>Error :(</p>;
+        if (error) {
+          console.log(error);
+          return <p>Error :(</p>;
+        }
 
         const { persons = {} } = data || {};
